+++ pages/Login/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { RouteComponentProps } from 'react-router-native';
-import { View, StyleSheet, Image, ViewStyle, TextStyle, TouchableOpacity, KeyboardAvoidingView, KeyboardAvoidingViewProps} from 'react-native';
+import { View, StyleSheet, ViewStyle, TextStyle, TouchableOpacity } from 'react-native';
 import { AppTheme, AppConstants } from '../../config/DefaultConfig';
 import useTheme from "../../hooks/useTheme";
 import ThemedText from '../../components/UI/ThemedText';
@@ -8,6 +8,7 @@ import useConstants from '../../hooks/useConstants';
 import Input from '../../components/Chat/Input';
 import microValidator from 'micro-validator' ;
 import { isIOS } from '../../utils';
+import AuthLayout from '../../components/Chat/AuthLayout';
 
 interface LoginField {
     username?: string;
@@ -40,12 +41,6 @@ const validate = (data: LoginField): ValidationError => {
     return errors
 }
 
-let keyboardAvoidingViewProps: KeyboardAvoidingViewProps = {}
-
-if (isIOS()) {
-    keyboardAvoidingViewProps.behavior = "position"
-}
-
 const Login: React.FunctionComponent<RouteComponentProps> = ({
     history
 }: RouteComponentProps) => {
@@ -76,37 +71,21 @@ const Login: React.FunctionComponent<RouteComponentProps> = ({
 
     return (
         <>
-            <View style={style.mainContainer}>
-                <KeyboardAvoidingView {...keyboardAvoidingViewProps} keyboardVerticalOffset={keyboardVerticalOffset} enabled>
-                    <View style={style.container}>
-                        <Image
-                            source={{ uri: constants.appLogo }}
-                            style={{width: 100, height: 100, borderRadius: 50}}
-                        />
-                    </View>
-                    <View style={[style.contentContainer ,{paddingTop: 10, paddingBottom: 30}]}>
-                        <ThemedText styleKey="textColor" style={style.nameStyle}>{constants.appName}</ThemedText>
-                    </View>
-                    <Input
-                        placeholder={constants.usernamePlacerHolder}
-                        onChangeText={onChangeUsername}
-                        value={username}
-                        errors={errors.username}
-                    />
-                    <Input
-                        placeholder={constants.passwordPlacerHolder}
-                        onChangeText={onChangePassword}
-                        value={password}
-                        secureTextEntry={true}
-                        errors={errors.password}
-                    />
-                    <View style={[style.container,{paddingTop: 50}]}>
-                        <TouchableOpacity onPress={goToChatList} style={[style.loginStyle, {borderColor: theme.lightBottomColor}]}>
-                            <ThemedText styleKey="textColor">{constants.loginButton}</ThemedText> 
-                        </TouchableOpacity> 
-                    </View>
-                </KeyboardAvoidingView>
-            </View>
+            <AuthLayout goToLocation={goToChatList}>
+                <Input
+                    placeholder={constants.usernamePlacerHolder}
+                    onChangeText={onChangeUsername}
+                    value={username}
+                    errors={errors.username}
+                />
+                <Input
+                    placeholder={constants.passwordPlacerHolder}
+                    onChangeText={onChangePassword}
+                    value={password}
+                    secureTextEntry={true}
+                    errors={errors.password}
+                />
+            </AuthLayout>
             <View style={style.topContainer}>
                 <ThemedText styleKey="lightTextColor">{constants.signupCheck}</ThemedText>
                 <TouchableOpacity onPress={goToSignup}>
