+++ pages/Signup/index.tsx
@@ -1,13 +1,10 @@
 import React, { useState } from 'react';
 import { RouteComponentProps } from 'react-router-native';
-import { View, StyleSheet, Image, ViewStyle, TextStyle, TouchableOpacity, KeyboardAvoidingView, KeyboardAvoidingViewProps} from 'react-native';
-import { AppTheme, AppConstants } from '../../config/DefaultConfig';
-import useTheme from "../../hooks/useTheme";
-import ThemedText from '../../components/UI/ThemedText';
+import { AppConstants } from '../../config/DefaultConfig';
 import useConstants from '../../hooks/useConstants';
 import Input from '../../components/Chat/Input';
+import AuthLayout from '../../components/Chat/AuthLayout';
 import microValidator from 'micro-validator' ;
-import { isIOS } from '../../utils';
 
 interface LoginField {
     name?: string;
@@ -55,12 +52,6 @@ const validate = (data: LoginField): ValidationError => {
     return errors
 }
 
-let keyboardAvoidingViewProps: KeyboardAvoidingViewProps = {}
-
-if (isIOS()) {
-    keyboardAvoidingViewProps.behavior = "position"
-}
-
 const Signup: React.FunctionComponent<RouteComponentProps> = ({
     history
 }: RouteComponentProps) => {
@@ -84,106 +75,36 @@ const Signup: React.FunctionComponent<RouteComponentProps> = ({
     }
 
     const constants: AppConstants = useConstants();
-    const theme: AppTheme = useTheme();
-    const keyboardVerticalOffset = isIOS ? 40 : 0;
 
     return (
-        <>
-            <View style={style.mainContainer}>
-                <KeyboardAvoidingView {...keyboardAvoidingViewProps} keyboardVerticalOffset={keyboardVerticalOffset} enabled>
-                    <View style={style.container}>
-                        <Image
-                            source={{ uri: constants.appLogo }}
-                            style={{width: 100, height: 100, borderRadius: 50}}
-                        />
-                    </View>
-                    <View style={[style.contentContainer ,{paddingTop: 10, paddingBottom: 30}]}>
-                        <ThemedText styleKey="textColor" style={style.nameStyle}>{constants.appName}</ThemedText>
-                    </View>
-                    <Input
-                        placeholder={constants.namePlaceholder}
-                        onChangeText={onChangeName}
-                        value={name}
-                        errors={errors.name}
-                    />
-                    <Input
-                        placeholder={constants.usernamePlacerHolder}
-                        onChangeText={onChangeUsername}
-                        value={username}
-                        errors={errors.username}
-                    />
-                    <Input
-                        placeholder={constants.emailPlacerHolder}
-                        onChangeText={onChangeEmail}
-                        value={email}
-                        errors={errors.email}
-                    />
-                    <Input
-                        placeholder={constants.passwordPlacerHolder}
-                        onChangeText={onChangePassword}
-                        value={password}
-                        secureTextEntry={true}
-                        errors={errors.password}
-                    />
-                    <View style={[style.container,{paddingTop: 50}]}>
-                        <TouchableOpacity onPress={goToLogin} style={[style.loginStyle, {borderColor: theme.lightBottomColor}]}>
-                            <ThemedText styleKey="textColor">{constants.signupButton}</ThemedText> 
-                        </TouchableOpacity> 
-                    </View>
-                </KeyboardAvoidingView>
-            </View>
-        </>
+        <AuthLayout goToLocation={goToLogin}>
+            <Input
+                placeholder={constants.namePlaceholder}
+                onChangeText={onChangeName}
+                value={name}
+                errors={errors.name}
+            />
+            <Input
+                placeholder={constants.usernamePlacerHolder}
+                onChangeText={onChangeUsername}
+                value={username}
+                errors={errors.username}
+            />
+            <Input
+                placeholder={constants.emailPlacerHolder}
+                onChangeText={onChangeEmail}
+                value={email}
+                errors={errors.email}
+            />
+            <Input
+                placeholder={constants.passwordPlacerHolder}
+                onChangeText={onChangePassword}
+                value={password}
+                secureTextEntry={true}
+                errors={errors.password}
+            />
+        </AuthLayout>
     );
 }
 
 export default Signup;
-
-interface Style {
-    mainContainer: ViewStyle;
-    container: ViewStyle;
-    contentContainer: ViewStyle;
-    nameStyle: TextStyle;
-    topContainer: ViewStyle;
-    loginStyle: ViewStyle;
-}
-
-const style: Style = StyleSheet.create<Style>({
-    mainContainer: {
-        flexDirection: 'column',
-        justifyContent: "center",
-        alignItems: 'stretch',
-        padding: 30,
-        flex: 1
-    },
-    container: {
-        flexDirection: 'row',
-        justifyContent: "center",
-        alignItems: 'center',
-        padding: 15,
-    },
-    contentContainer: {
-        flexDirection: 'row',
-        justifyContent: "center",
-        paddingTop: 10,
-        paddingBottom: 10,
-    },
-    nameStyle: {
-        fontWeight: "bold",
-        fontSize: 24,
-    },
-    topContainer: {
-        flexDirection: 'column',
-        justifyContent: "flex-end",
-        alignItems: 'center',
-        paddingBottom: 10
-    },
-    loginStyle: {
-        flexDirection: 'row',
-        justifyContent: "center",
-        alignItems: 'center',
-        width: 170,
-        borderWidth: 2,
-        borderRadius: 50,
-        padding: 10,
-    }
-})
